add types file, move things around
