+++ src/commands/list_dependencies.py
@@ -3,13 +3,8 @@ from __future__ import annotations
 from argparse import _SubParsersAction
 from argparse import ArgumentParser
 from argparse import Namespace
-from typing import Optional
 
 from configs.service_config import load_service_config
-from configs.service_config import ServiceConfig
-from exceptions import ConfigError
-from exceptions import ServiceNotFoundError
-from services import find_matching_service
 
 
 def add_parser(subparsers: _SubParsersAction[ArgumentParser]) -> None:
@@ -29,19 +24,7 @@ def list_dependencies(args: Namespace) -> None:
     """List the dependencies of a service."""
     service_name = args.service_name
 
-    service_config: Optional[ServiceConfig] = None
-    if service_name is not None:
-        try:
-            service_config = find_matching_service(service_name).service_config
-        except ServiceNotFoundError as e:
-            print(e)
-            return
-    else:
-        try:
-            service_config = load_service_config()
-        except ConfigError as e:
-            print(e)
-            return
+    service_config = load_service_config(service_name)
 
     dependencies = service_config.dependencies
