+++ src/configs/service_config.py
@@ -1,52 +1,17 @@
 from __future__ import annotations
 
 import os
-from dataclasses import dataclass
-from typing import Dict
-from typing import List
-from typing import Optional
 
 import yaml
+from configs.types import Dependency
+from configs.types import ServiceConfig
 from constants import DEVSERVICES_DIR_NAME
 from constants import DOCKER_COMPOSE_FILE_NAME
+from exceptions import ConfigError
 from exceptions import ConfigNotFoundError
 from exceptions import ConfigParseError
-from exceptions import ConfigValidationError
-
-
-@dataclass
-class Dependency:
-    description: str
-    link: Optional[str] = None
-
-
-@dataclass
-class ServiceConfig:
-    version: float
-    service_name: str
-    dependencies: Dict[str, Dependency]
-    modes: Dict[str, List[str]]
-
-    def __post_init__(self) -> None:
-        self._validate()
-
-    def _validate(self) -> None:
-        if self.version != 0.1:
-            raise ConfigValidationError(
-                f"Invalid version '{self.version}' in service config"
-            )
-
-        for mode, services in self.modes.items():
-            for service in services:
-                if service not in self.dependencies:
-                    raise ConfigValidationError(
-                        f"Service '{service}' in mode '{mode}' is not defined in dependencies"
-                    )
-
-
-@dataclass
-class Config:
-    service_config: ServiceConfig
+from exceptions import ServiceNotFoundError
+from utils.services import find_matching_service
 
 
 def load_service_config_from_file(repo_path: str) -> ServiceConfig:
@@ -54,9 +19,7 @@ def load_service_config_from_file(repo_path: str) -> ServiceConfig:
         repo_path, DEVSERVICES_DIR_NAME, DOCKER_COMPOSE_FILE_NAME
     )
     if not os.path.exists(config_path):
-        raise ConfigNotFoundError(
-            f"Config file not found in current directory: {config_path}"
-        )
+        raise ConfigNotFoundError(f"Config file not found in directory: {config_path}")
     with open(config_path, "r") as stream:
         try:
             config = yaml.safe_load(stream)
@@ -83,6 +46,22 @@ def load_service_config_from_file(repo_path: str) -> ServiceConfig:
             ) from yml_error
 
 
-def load_service_config() -> ServiceConfig:
+def load_service_config_from_current_directory() -> ServiceConfig:
     """Load the service config for the current directory."""
     return load_service_config_from_file(os.getcwd())
+
+
+def load_service_config(service_name: str | None) -> ServiceConfig:
+    if service_name is not None:
+        try:
+            service_config = find_matching_service(service_name).service_config
+        except ServiceNotFoundError as e:
+            print(e)
+            return
+    else:
+        try:
+            service_config = load_service_config_from_current_directory()
+        except ConfigError as e:
+            print(e)
+            return
+    return service_config
